Honestly it's been a long time since I uploaded and I don't remember all the different updates that were made since the last commit.
In `Solver.py` it looks like some constraint logic was moved from `solving_pass` into `solve_line`. I believe it serves an additional purpose now too but I'm not totally clear what the change is.

I remember this being some two-birds-with-one-stone code, and I think that's because it was also able to replace a massive portion of the code in `ClueRun.solve_self`, but the `solve_self` code ended up speeding up the algorithm considerably still, so I uncommented it again.

I know there was also a bug in the code at one point that I needed to sort out, so perhaps the replaced logic in `solving_pass` was not a full replacement for the commented out code in `ClueRun.solve_self` and this new code is what was able to actually cover all the functionality needed. In any case, I know this is the best version of the code yet!

I'm also inlcuding some piskel files for drawings that helped me at one point or another. I don't remember what's in them.